fixed issues, added ui package
