+++ app/src/main/java/com/example/movieapp/ui/movies/MoviesFragmentViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.movieapp.movies
+package com.example.movieapp.ui.movies
 
 import android.content.Context
 import androidx.lifecycle.LiveData
@@ -9,6 +9,7 @@ import com.example.movieapp.api.ApiErrorHandling
 import com.example.movieapp.api.RetrofitRepository
 import com.example.movieapp.database.DatabaseRepository
 import com.example.movieapp.models.Movie
+import com.example.movieapp.models.MovieDetails
 import com.example.movieapp.utils.Constants.POPULAR
 import com.example.movieapp.utils.Constants.SAVED
 import com.example.movieapp.utils.Constants.TOP_RATED
@@ -28,8 +29,8 @@ class MoviesFragmentViewModel @Inject constructor(private val retrofitRepository
 
     private var _livedata = MutableLiveData<ApiErrorHandling<Movie>>()
     var livedata: LiveData<ApiErrorHandling<Movie>> = _livedata
-    private var _savedLiveData = MutableLiveData<MutableList<Movie.Results>>()
-    var savedLiveData: LiveData<MutableList<Movie.Results>> = _savedLiveData
+    private var _savedLiveData = MutableLiveData<MutableList<MovieDetails>>()
+    var savedLiveData: LiveData<MutableList<MovieDetails>> = _savedLiveData
     private var _isTopRated = MutableLiveData<Boolean>()
     var isTopRated : LiveData<Boolean> = _isTopRated
     private val _isNetworkAvailable = MutableLiveData<Boolean?>(null)
@@ -62,9 +63,9 @@ class MoviesFragmentViewModel @Inject constructor(private val retrofitRepository
         viewModelScope.launch {
             withContext(Dispatchers.IO) {
                 val result = databaseRepository.getSavedMovies()
-                val favoriteMovies: MutableList<Movie.Results> = mutableListOf()
+                val favoriteMovies: MutableList<MovieDetails> = mutableListOf()
                 result.forEach {
-                    favoriteMovies.add(Movie.Results(it.poster, it.id))
+                    favoriteMovies.add(MovieDetails(it.poster, it.id))
                 }
                 _savedLiveData.postValue(favoriteMovies)
             }
