+++ app/src/main/res/layout/fragment_splash_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/bgr_color"
-    tools:context=".splash.SplashScreen">
+    tools:context=".ui.splash.SplashScreen">
 
 
     <TextView
