+++ app/src/main/res/layout/movies_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".movies.MoviesFragment"
+    tools:context=".ui.movies.MoviesFragment"
     android:background="@color/bgr_color">
